Fix testing setup
The testing had a lot of issues around libstdc++ messing this up when
we try to load the tree sitter grammers. What I have found is that the
c ts lang module is dynamically linked. So we can use that without any
issues.
Refer to this commit for more info:
https://github.com/emacs-tree-sitter/tree-sitter-langs/commit/a01730c87235d8ec323dc70bf49d753f81087607